refactor(website): render both nav panel surfaces from one body
The desktop hover panel and the mobile drill-in stack each wrote the same
shape by hand: map panel.columns to a column wrapper, map column.items
through NavPanelItem, then render panel.footer with nav-panel-footer /
nav-panel-footer-lead plus one more NavPanelItem. Individual items already
funnelled through the shared NavPanelItem, which is what has kept the
analytics ids from drifting, but the column and footer shape around them
was duplicated and free to diverge.

NavPanelBody now renders columns + footer once, parameterised by the two
things that genuinely differ: the wrapper class names (the layouts are not
the same, so desktop keeps nav-panel-cols / nav-panel-col and mobile keeps
nav-mobile-panel / nav-mobile-group) and the optional onNavigate the mobile
drawer uses to close itself. The caller still owns the outermost element,
because that is the other real difference: desktop needs the panel id and
data-columns, mobile needs neither. Desktop passes columnsClassName, mobile
omits it, so the mobile stack keeps its columns as direct children and the
scoped .nav-mobile-panel .nav-panel-footer override still matches.

NavPanelItem and trackNavItem move into the new module with it. NavMobile
had been importing NavPanelItem from NavDesktop, which pointed the
dependency the wrong way between two sibling surfaces; both now import from
a shared leaf and neither imports the other.

Pure refactor — no rendered output changes. Verified by dumping the
outerHTML of all three panels on both surfaces before and after: identical
byte for byte, down to the useId-generated panel id. nx test website 1462
passed, nx build website green, and the nav e2e suites (nav-panels,
nav-drawer, nav-surface, nav-height) 32 passed in a real browser.
